Consolidate music discovery into music page
